% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_for_missing.R
\name{check_for_missing}
\alias{check_for_missing}
\title{Check for Missing Values}
\usage{
check_for_missing(x, colnames = NULL, error = TRUE)
}
\arguments{
\item{x}{a data.frame}

\item{colnames}{names in the data.frame that should not contain missing values.
If \code{NULL} (default) all columns are considered}

\item{error}{default \code{TRUE}, prints a table of columns with the number of
missing values to the console and returns an error if any missing values are
found. If \code{FALSE} and missing values are found, a warning is displayed
and the columns with number of missing are return as a data.frame. If no
missing values are found, a success message is printed returns \code{NULL}}
}
\description{
Identifies whether missing values are present in a data.frame.
Default behaviour will return an error if any missing values are present
from select columns.
}
