% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_excel_date.R
\name{clean_excel_date}
\alias{clean_excel_date}
\title{Clean up excel date field}
\usage{
clean_excel_date(x, time = FALSE, verbose = FALSE)
}
\arguments{
\item{x}{a string to be formatted as a date}

\item{time}{default \code{FALSE}, whether to return a dateime instead of date}

\item{verbose}{default \code{FALSE} hides any warning messages from \code{janitor::excel_numeric_to_date()} and/or \code{lubridate::parse_date_time()}}
}
\value{
a \code{Date} when \code{time=FALSE} or a \code{POSIXct} when \code{time=TRUE}
}
\description{
Sometimes excel dates are strings, sometimes 5 digit codes, sometimes both
}
\examples{
    clean_excel_date("32545")
    clean_excel_date("12/2/2021")
    clean_excel_date("6/28/2021")

}
