% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_numbers.R
\name{clean_numbers}
\alias{clean_numbers}
\title{Clean Numbers}
\usage{
clean_numbers(x, digits = 2, delim = ",")
}
\arguments{
\item{x}{object to be parsed}

\item{digits}{the digits to which to round, default of \code{2}}

\item{delim}{the delimiter to use for the big.mark arguments
default of \code{","}}
}
\value{
a number formatted as specified (e.g. 1000 becomes 1,000) with defaults
}
\description{
Clean Numbers
}
\examples{
clean_numbers(10000)
}
