% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counter_verbose.R
\name{counter_verbose}
\alias{counter_verbose}
\title{Verbose Counter Completion in Loops}
\usage{
counter_verbose(counter_value, counter_length)
}
\arguments{
\item{counter_value}{number of times the for loop has gone through}

\item{counter_length}{length of the for loop}
}
\description{
In a for loop, print the (rough) percent of loops that are
finished.
}
\examples{
for(i in 1:28){
  cat("\n\ti = ",i)
  counter_verbose(i,28)
}
}
