% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-combined-plate.R
\name{extract_combined_plate}
\alias{extract_combined_plate}
\title{Extract combined plate number from a character string}
\usage{
extract_combined_plate(input)
}
\arguments{
\item{input}{A character string containing a combined plate number}
}
\value{
The combined plate number as a string
}
\description{
\code{extract_combined_plate} extracts the combined plate number from an input
string, using \code{extract_id} and \code{regex_ids}. A full description of the
format of the combined plate identifier is included in the documentation
for \code{regex_ids}.
}
\examples{
extract_combined_plate("WS123456_CP12345_results.csv")

}
