% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-id.R
\name{extract_id}
\alias{extract_id}
\title{Extract an identifier using regex}
\usage{
extract_id(input, pattern, group)
}
\arguments{
\item{input}{The string to search}

\item{pattern}{The regular expression to match}

\item{group}{The group within the regular expression}
}
\value{
The matching identifier from the string
}
\description{
\code{extract_id} is a generic function for extracting a pattern from strings,
with input and output checks.
It is used as the basis for specific \verb{extract_*} functions such as
\code{extract_labno} and \code{extract_worksheet} to reduce code
duplication.
}
\examples{

# Example to find labno
file_string <- "WS123456_12345678_sequencing_results.xlsx"

extract_id(file_string,
           pattern = regex_ids()$labno_suffix$regex,
           group = regex_ids()$labno_suffix$labno_group)

# Errors when empty string supplied
try(extract_id("",
             pattern = regex_ids()$labno_suffix$regex,
             group = regex_ids()$labno_suffix$labno_group))

}
