% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-igene-dno.R
\name{extract_igene_dno}
\alias{extract_igene_dno}
\title{Extract iGene D number from a character string}
\usage{
extract_igene_dno(input)
}
\arguments{
\item{input}{A character string containing an iGene D number}
}
\value{
The iGene D number as a string
}
\description{
\code{extract_igene_dno} extracts the iGene derivative number from a string,
using \code{extract_ids}. A description of the format of iGene referral numbers
is included in the documentation for \code{regex_ids}.
}
\examples{
extract_igene_dno("MPL-00001_D26-AB12_ctDNA_M1_T26-AB12")

}
