% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-igene-rno.R
\name{extract_igene_rno}
\alias{extract_igene_rno}
\title{Extract iGene R number from a character string}
\usage{
extract_igene_rno(input)
}
\arguments{
\item{input}{A character string containing an iGene R number}
}
\value{
The iGene R number as a string
}
\description{
\code{extract_igene_rno} extracts the iGene referral number from a string,
using \code{extract_ids}. A description of the format of iGene referral numbers
is included in the documentation for \code{regex_ids}.
}
\examples{
extract_igene_rno("WS123456_R26-12AB")

}
