% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-igene-sno.R
\name{extract_igene_sno}
\alias{extract_igene_sno}
\title{Extract iGene S number from a character string}
\usage{
extract_igene_sno(input)
}
\arguments{
\item{input}{A character string containing an iGene S number}
}
\value{
The iGene S number as a string
}
\description{
\code{extract_igene_sno} extracts the iGene sample number from a string, using
\code{extract_id} and \code{regex_ids}. A description of the format of iGene sample
numbers is included in the documentation for \code{regex_ids}.
}
\examples{
extract_igene_sno("WS123456_S26-12AB")

}
