% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_igene-tno.R
\name{extract_igene_tno}
\alias{extract_igene_tno}
\title{Extract iGene T number from a character string}
\usage{
extract_igene_tno(input)
}
\arguments{
\item{input}{A character string containing an iGene T number}
}
\value{
The iGene T number as a string
}
\description{
\code{extract_igene_tno} extracts the iGene test number from a string,
using \code{extract_ids}. A description of the format of iGene referral numbers
is included in the documentation for \code{regex_ids}.
}
\examples{
extract_igene_tno("MPL-00001_D26-AB12_ctDNA_M1_T26-AB12")

}
