% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-labno.R
\name{extract_labno}
\alias{extract_labno}
\title{Extract lab number from a string}
\usage{
extract_labno(input)
}
\arguments{
\item{input}{A character string containing a lab number}
}
\value{
The lab number as a string
}
\description{
\code{extract_labno} extracts the DNA lab number (labno) from an input string,
using \code{extract_id} and \code{regex_ids}. A full description of the format of
the labno identifier is included in the documentation for \code{regex_ids}.
}
\examples{
extract_labno("WS123456_12345678")

}
