% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-suffix.R
\name{extract_suffix}
\alias{extract_suffix}
\title{Extract a sample replicate suffix from a string}
\usage{
extract_suffix(input)
}
\arguments{
\item{input}{A character string containing a lab number and a suffix}
}
\value{
A character string of the suffix. Inputs without a suffix with
return "".
}
\description{
\code{extract_suffix} extracts the letter suffix from after a DNA lab number.
Letter suffixes are used when samples are tested in multiple replicates
(Examples: 12345678a, 12345678b, 12345678c).
More detail is provided in the documentation for \code{regex_ids}.
}
\examples{
# Example with suffix
extract_suffix("WS123456_12345678a_replicate1")

# Empty string returned when no suffix is present
extract_suffix("WS123456_12345678_replicate1")

}
