% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-wgs-patient-no.R
\name{extract_wgs_patient_no}
\alias{extract_wgs_patient_no}
\title{Extract WGS patient number from a character string}
\usage{
extract_wgs_patient_no(input)
}
\arguments{
\item{input}{A character string containing a WGS patient number}
}
\value{
The WGS patient number as a string
}
\description{
\code{extract_wgs_patient_no} extracts the whole genome sequencing (WGS)
patient number from an input string, using \code{extract_id} and \code{regex_ids}.
A description of the format of the WGS patient number is included in the
documentation for \code{regex_ids}.
}
\examples{
extract_wgs_patient_no("1234567890_p12345678901_LP1234567-DNA-reported_variants.csv")

}
