% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-wgs-referral-no.R
\name{extract_wgs_referral_no}
\alias{extract_wgs_referral_no}
\title{Extract WGS referral number from a character string}
\usage{
extract_wgs_referral_no(input)
}
\arguments{
\item{input}{A character string containing a WGS referral number}
}
\value{
The WGS referral number as a string
}
\description{
\code{extract_wgs_referral_no} extracts the whole genome sequencing (WGS)
referral number from an input string, using \code{extract_id} and \code{regex_ids}.
A description of the format of the WGS referral number is included in the
documentation for \code{regex_ids}.
}
\examples{
extract_wgs_referral_no("Referral ID: r12345678901")

}
