% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-worksheet.R
\name{extract_worksheet}
\alias{extract_worksheet}
\title{Extract worksheet from a character string}
\usage{
extract_worksheet(input)
}
\arguments{
\item{input}{A character string containing a worksheet identifier}
}
\value{
The worksheet identifier as a string
}
\description{
\code{extract_worksheet} extracts the worksheet identifier from an input string,
relying on \code{extract_id} and \code{regex_ids}.
Details of the worksheet number format are given in the documentation for
\code{regex_ids}.
}
\examples{
extract_worksheet("WS123456_12345678")

}
