% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fclip.R
\name{fclip}
\alias{fclip}
\title{Truncation of Values}
\usage{
fclip(x, xmin = NULL, xmax = NULL)
}
\arguments{
\item{x}{a vector of values}

\item{xmin}{a double representing the minimum bound}

\item{xmax}{a double representing the upper bound}
}
\description{
Truncation of Values
}
\examples{
x <- rnorm(100, 1,2)

hist(x)

hist(fclip(x, 0, 3))


}
\author{
\url{https://github.com/rmcelreath/rethinking/blob/master/R/utilities.r}
}
