% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shorthand.R
\name{g}
\alias{g}
\title{Shorthand for dplyr::glimpse}
\usage{
g(...)
}
\arguments{
\item{...}{parameters passed to \code{glimpse()}}
}
\value{
x original x is (invisibly) returned, allowing \code{glimpse()} to be used within a data pipe line.
}
\description{
see \link[dplyr]{glimpse} documentation
}
