% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_template.R
\name{generate_template}
\alias{generate_template}
\title{Generate Template File}
\usage{
generate_template(
  template_path,
  glue_open = "{",
  glue_close = "}",
  envir = parent.frame()
)
}
\arguments{
\item{template_path}{the path to a file}

\item{glue_open}{default \code{\{}, the opening delimiter}

\item{glue_close}{default \code{\}}, the closing delimiter}

\item{envir}{environment to evaluate the \code{glue} expression. default
is \code{parent.frame()}.}
}
\value{
a character string of file contents
}
\description{
this function reads in most types of text/code-based files
(e.g. \code{.sql}, \code{.txt}, \code{.R}) and collapses them into a single character string.
supports \code{glue::glue} syntax within the file.
}
