% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dups.R
\name{get_dups}
\alias{get_dups}
\title{Get Duplicates from Vector or Data Frame}
\usage{
get_dups(x, data = NULL)
}
\arguments{
\item{x}{A vector to get duplicate data from. Alternatively, the \strong{quoted}
column name to return duplicates from when \code{data} is provided.}

\item{data}{A data frame to get duplicate data from. Defaults \code{NULL}.}
}
\description{
Find and return all duplicates (not just the first indices like
\code{\link[base]{duplicated}} does)
}
\examples{
# vector example
a <- c(1,2,3,4,4,5,6,7,7,8,9,10)
get_dups(x = a)

#data.frame example
b <- data.frame(id = a, counter = seq_along(a))
get_dups(x = 'id', data = b)
}
