% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shorthand.R
\name{gg}
\alias{gg}
\title{Shorthand for glue::glue}
\usage{
gg(...)
}
\arguments{
\item{...}{parameters passed to \code{glue()}}
}
\value{
a glue character string
}
\description{
see \link[glue]{glue} documentation
}
