% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handleTime.R
\name{handleTime}
\alias{handleTime}
\title{Add Time to R Console Handle}
\usage{
handleTime(time = T, format = "\%m/\%d \%r")
}
\arguments{
\item{time}{if \code{TRUE} (default), will put the date/time in the R Console
handle, \code{FALSE} will revert to the standard \code{> } handle.}

\item{format}{desired format of the timestamp. defaults to \code{\%m/\%d \%r}}
}
\description{
Add the time to the console for when you want to casually time
your code but down want to wrap lots of code with \code{\link[base]{system.time}}
or intermittently check the \code{\link[base]{Sys.time}}.
}
\examples{
# turn it on:
handleTime()

#turn if off:
handleTime(FALSE)

}
