% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shorthand.R
\name{hh}
\alias{hh}
\title{Shorthand for here::here}
\usage{
hh(...)
}
\arguments{
\item{...}{parameters passed to \code{here()}}
}
\value{
root path to project plus any additional path components provided
}
\description{
see \link[here]{here} documentation
}
