% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_to.R
\name{inf_to}
\alias{inf_to}
\title{Replace infinite values}
\usage{
inf_to(x, to = 0)
}
\arguments{
\item{x}{a vector}

\item{to}{value of length 1 to change \code{Inf} and \code{-Inf} values to, or a vector of same length as x where any \code{Inf}/\code{-Inf} values in x will be replaced with the corresponding values in the 'to' vector. Default \code{0}.}
}
\description{
Allows for replacement of \code{Inf} and \code{-Inf} values in a vector.
}
\seealso{
\code{\link[=na_to]{na_to()}}, \code{\link[=nan_to]{nan_to()}}, \code{\link[=nonfinite_to]{nonfinite_to()}}
}
