% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_perc.R
\name{make_perc}
\alias{make_perc}
\title{Make a Percentage}
\usage{
make_perc(x, digitz = 0, type = "numeric")
}
\arguments{
\item{x}{the value to be converted}

\item{digitz}{the number of digits desired with default of 0
(e.g. .95111 = 95 for a value of 0 digits)}

\item{type}{do you want it to be a character or number}
}
\description{
Make a Percentage
}
