% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_proper_list.R
\name{make_proper_list}
\alias{make_proper_list}
\title{A function to Make Proper Lists}
\usage{
make_proper_list(x, conjunction = "and")
}
\arguments{
\item{x}{the vector of items}

\item{conjunction}{the conjunction to use with default \code{"and"}}
}
\description{
A function to Make Proper Lists
}
\examples{
make_proper_list("Apples")
make_proper_list(c("Apples", "Bananas"))
make_proper_list(c("Apples", "Bananas", "Pears"))
}
