% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate-ids.R
\name{mutate_ids}
\alias{mutate_ids}
\title{Add sample identifiers from a filename as additional data columns}
\usage{
mutate_ids(df, id_col = filename)
}
\arguments{
\item{df}{The dataframe to add identifiers to}

\item{id_col}{The column in the dataframe which includes the filename.
Defaults to "filename".}
}
\value{
The original dataframe with additional columns of identifiers
from the filename column
}
\description{
\code{mutate_ids} extracts the three most common sample identifiers (worksheet,
labno and suffix) from a column
within a dataframe, and adds them as new separate columns with
\code{dplyr::mutate}.
Details on the formats of each identifier are included in the documentation
for \code{regex_ids}.

This function is particularly effective when combined with the pipe (|>)
operator as it takes a dataframe as its input.
Extra detail of using \code{mutate_ids} is provided in the vignette.
}
\examples{

# Dataframe of results from multiple files with filename as a column
results_df <- tibble::tribble(
~result, ~filename,
"Variant detected", "WS123456_12345678a_PierreBEZUKHOV.xlsx",
"Variant not detected", "WS123456_23456789_AnnaKARENINA.xlsx",
"Variant not detected", "WS123456_34567890_IvanILYICH.xlsx")

# Use `mutate_ids` to separate identifiers
results_df |>
  mutate_ids()

}
