% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_to.R
\name{na_to}
\alias{na_to}
\title{Replace NAs}
\usage{
na_to(x, to = 0)
}
\arguments{
\item{x}{a vector}

\item{to}{value of length 1 to change \code{NA} values to, or a vector of same length as x where any \code{NA} values in x will be replaced with the corresponding values in the 'to' vector. Default \code{0}.}
}
\description{
Allows for replacement of \code{NA} values in a vector.

Unlike \code{dplyr::replace_na()}, the replacement can be a vector (i.e. works within \code{dplyr::mutate} where \code{NA} values can be replaced with values in another column).
}
\seealso{
\code{\link[=nan_to]{nan_to()}}, \code{\link[=inf_to]{inf_to()}}, \code{\link[=nonfinite_to]{nonfinite_to()}}
}
