% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_to.R
\name{nan_to}
\alias{nan_to}
\title{Replace NaNs}
\usage{
nan_to(x, to = 0)
}
\arguments{
\item{x}{a vector}

\item{to}{value of length 1 to change \code{NaN} values to, or a vector of same length as x where any \code{Nan} values in x will be replaced with the corresponding values in the 'to' vector. Default \code{0}.}
}
\description{
Allows for replacement of \code{NaN} values in a vector.
}
\seealso{
\code{\link[=na_to]{na_to()}}, \code{\link[=inf_to]{inf_to()}}, \code{\link[=nonfinite_to]{nonfinite_to()}}
}
