% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_to.R
\name{nonfinite_to}
\alias{nonfinite_to}
\title{Replace all non-finite values}
\usage{
nonfinite_to(x, to = 0)
}
\arguments{
\item{x}{a vector}

\item{to}{value of length 1 to change non-finite values to, or a vector of same length as x where any non-finite values in x will be replaced with the corresponding values in the 'to' vector. Default \code{0}.}
}
\description{
Allows for replacement of non-finite values in a vector.
This is a shortcut for replacing \code{NA}, \code{NaN}, \code{Inf}, and \code{-Inf} values with some other value.
}
\seealso{
\code{\link[=na_to]{na_to()}}, \code{\link[=nan_to]{nan_to()}}, \code{\link[=inf_to]{inf_to()}}
}
