% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perc_change.R
\name{perc_change}
\alias{perc_change}
\title{Percent Change Calculations}
\usage{
perc_change(start, end, p = NULL, returnall = FALSE)
}
\arguments{
\item{start}{starting value}

\item{end}{ending value}

\item{p}{percent change}

\item{returnall}{whether or not to return all arguments. defaults to \code{FALSE} and only returns the NULL parameter}
}
\description{
Compute the percent change between two numbers or determine parameters to obtain a certain percent change.

Like \code{stats.power.xxx.test()} functions, exactly one argument should be null and that parameter is calculated
}
\examples{
perc_change(0.75, .375) #50\% decrease
perc_change(0.75, end = NULL, p = -.5, returnall = TRUE) #50\% decrease

}
