% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_mrn.R
\name{safe_pad}
\alias{safe_pad}
\title{Safely pad a vector to the same length}
\usage{
safe_pad(x, length = NULL, pad_with = "0", pad_n = NULL, side = "left")
}
\arguments{
\item{x}{a vector of strings or numbers to be padded}

\item{length}{length of the string to return; default of \code{NULL} uses the longest character string in the vector}

\item{pad_with}{one character, default = \code{'0'}}

\item{pad_n}{if supplied, length is ignored. This is the number of 0s to pad on top of the longest character string. Example: if pad_n = 1 and the longest string contains 5 characters, all strings will be padded to a length of 5+1 = 6 characters.}

\item{side}{which side of the string to pad with zero. default = \code{left}}
}
\description{
formats a value with the same leading or trailing value
}
\examples{
safe_pad(c(123,123456,1234567))
safe_pad(c(123,123456,1234567), 9)
safe_pad(c(123,123456,1234567), pad_n = 1)
safe_pad(c(123,123456,1234567), pad_with = ' ')
}
