% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{safe_pkg_install}
\alias{safe_pkg_install}
\title{safe installation of packages}
\usage{
safe_pkg_install(x, repo = c(CRAN = "http://cran.rstudio.com"))
}
\arguments{
\item{x}{the package you wish to install}

\item{repo}{the repo you wish to use}
}
\description{
This function checks if a package exists, if it does it installs
it. If it does not find the package in CRAN it will let you know
}
