% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setwd_currentfile.R
\name{setwd_currentfile}
\alias{setwd_currentfile}
\title{Set Working Directory to the Path of the Current Script}
\usage{
setwd_currentfile()
}
\description{
Changes the current working directory to the Current Script in
view in the source editor.
}
