% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{suppress_all}
\alias{suppress_all}
\title{Suppress all messages and warnings}
\usage{
suppress_all(x)
}
\arguments{
\item{x}{the item passed to the function, generally a function}
}
\description{
Suppress all messages and warnings
}
