% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_xlsx.R
\name{to_xlsx}
\alias{to_xlsx}
\title{Write dataframe to excel}
\usage{
to_xlsx(
  df,
  file,
  sheet = "Sheet1",
  overwrite = TRUE,
  table_style = "TableStyleMedium16",
  max_col_width = 22,
  ...
)
}
\arguments{
\item{df}{a dataframe}

\item{file}{A character string naming an xlsx file}

\item{sheet}{A name for the worksheet}

\item{overwrite}{If \code{TRUE}, overwrite any existing file.}

\item{table_style}{Any excel table style name or "none" (see "formatting" vignette in \code{openxlsx} package).}

\item{max_col_width}{maximum column width allowed, default \code{22}}

\item{...}{other parameters passed to \code{openxlsx::writeDataTable}}
}
\description{
Writes a dataframe to an xlsx using \code{openxlsx::saveWorkbook}
}
