% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_with_meta.R
\name{write_with_meta}
\alias{write_with_meta}
\title{Write CSV with Meta Information}
\usage{
write_with_meta(x, filepath, .max_n = 10)
}
\arguments{
\item{x}{a data.frame to be written}

\item{filepath}{a string representing the filepath of the file}

\item{.max_n}{an integer indicating the maximum number of
unique values to print}
}
\description{
Writes a corresponding README file with some summary
information about the csv. Uses d\code{ata.table::fread} to
write the actual file. Only supports those outputs
available to \code{data.table::fread}
}
\examples{

tmp <- tempfile(fileext = ".csv")
out <- write_with_meta(warpbreaks, tmp)

writeLines(readLines(out))
}
