% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex-ids.R
\name{regex_ids}
\alias{regex_ids}
\title{Regular expressions for sample identifiers}
\usage{
regex_ids()
}
\value{
A named list of regular expressions for different identifiers
}
\description{
This function is a centralised place for regular expressions for extracting
different identitiers.
Regular expressions are written as raw strings using \code{stringr::regex} which
makes them easier to understand and maintain.
The \href{https://r4ds.hadley.nz/regexps.html}{"Regular Expressions" chapter in "R for Data Science"}
is a great resource for understanding regular expressions.
}
\examples{
regex_ids()$labno_suffix
}
