% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex-ids.R
\name{regex_ids}
\alias{regex_ids}
\title{Regular expressions for sample identifiers}
\usage{
regex_ids()
}
\value{
A named list of regular expressions for different identifiers
}
\description{
\code{regex_ids} is a centralised place for regular expressions (regex) for extracting
different identifiers. These expressions are then used by the
\verb{extract_*} series of functions for extracting identifiers.

Regular expressions are written as raw strings using \code{stringr::regex} which
makes them easier to understand and maintain.
The \href{https://r4ds.hadley.nz/regexps.html}{"Regular Expressions" chapter in "R for Data Science"}
is a great resource for understanding regular expressions.
}
\section{DNA Database identifiers}{


\strong{worksheet}: the worksheet identifier is "WS" followed by six numbers, which
increase sequentially (i.e "WS000001", "WS000002", WS000003"). Worksheets
are the way in which genetic tests are organised into batches within the
lab. Worksheet identifiers are generated by DNA Database.

\strong{combined_plate}: the combined plate number from the DNA Database which is
"CP" followed by 5 digits. "Combined plates" are created by combining
muliple worksheets together.

\strong{labno}: the numeric identifier given to every DNA sample from DNA
Database. The regex specifies 8 digits in the identifier, although for
samples from 2012 or earlier there will only be 6 digits. This identifier
is referred to as "labno" throughout \code{idtools} because this is the name
used in the "Samples" table of DNA Database.

The first two digits of the labno format are the year the sample was received,
and then the digits increase with each new sample. For example, the first
sample received in 2026 would have the labno 26000001.

\strong{suffix}: if the same DNA sample is tested multiple times on the same
worksheet (which can happen
for repeatability experiments during test validation) then a suffix letter
(usually a-d) is added directly after the labno. Extracting this suffix is
useful when assessing the variability between different replicates of the
same sample.
}

\section{iGene identifiers}{


\strong{igene_rno}: the R (referral) number from the iGene database has the format
of R followed by the two-digit year number, a hyphen, and then four
alpha-numeric charactes. Example: R26-AB12.

\strong{igene_sno}: the S (sample) number from the iGene database has the same
format as the R number (above) but with an S instead of an R a the
beginning.
}

\section{Whole genome sequencing identifiers}{


\strong{wgs_patient_no}: the whole genome sequencing (WGS) pathway in the Genomic
Medicine Service uses a patient identifier which is "p" followed by 11 digits.

\strong{wgs_referral_no}: the referral ("r") equivalent of the WGS patient number,
structured as "r" followed by 11 digits.
}

\examples{

file_string <- "WS123456_12345678_sequencing_results.xlsx"

stringr::str_extract(file_string,
                      pattern = regex_ids()$labno_suffix$regex,
                      group = regex_ids()$labno_suffix$labno_group)

}
